feat: rename package check
